Move RPC classes into core
